feat: dc_motor TestSave, add some TODO

+++ common/Inc/myints.h
@@ -12,7 +12,7 @@ typedef uint32_t u64;
 typedef int8_t i8;
 typedef int16_t i16;
 typedef int32_t i32;
-typedef int32_t i64;
+typedef int64_t i64;
 #endif
 
 #ifdef USE_FXX
